Link author nicks (#315)
* Remove trailing whitespace

* Fix grammar

* Remove multiple spaces

* Link author nicks

Author nicks were already rendered in blue text color exactly like
links. However, being text only, they could not be clicked, which was
quite a bad user experience. So visitors needed to figure out that
avatars are the actual links, which they could follow to end up on the
user profile. The problem is, that this is not immediately obvious. This
commit turns the author nicks into proper links.

It also fixes a broken private message sender link.

I kept the `p-name` CSS class, since it specifies a slightly larger font
size compared to `u-url`. I don't care about those specifics at all, so
I'm open to discussions.

* Fix checkbox label

The checkbox is actually named `agree` rather than `switch`, so the
label did not work and one had to click on the small slider thingy in
order to (un)check it. Since the label encloses the input anyways, its
`for` attribute is not needed at all.

* Regenerate rice-box by `make generate`